Class test skeleton set-up
