+++ src/Hotel.js
@@ -1,0 +1,7 @@
+class Hotel {
+  constructor() {
+
+  }
+}
+
+export default Hotel;
