+++ src/Rooms.js
@@ -1,0 +1,7 @@
+class Rooms {
+  constructor() {
+
+  }
+}
+
+export default Rooms;
