+++ src/User.js
@@ -1,0 +1,7 @@
+class User {
+  constructor() {
+
+  }
+}
+
+export default User;
